Moved stub-related classes to new Models.Stubs namespace
