+++ MbDotNet/MbDotNet.csproj
@@ -50,7 +50,7 @@
     <Compile Include="Enums\Protocol.cs" />
     <Compile Include="Exceptions\MountebankException.cs" />
     <Compile Include="Interfaces\IRequestProxy.cs" />
-    <Compile Include="Models\HttpStub.cs" />
+    <Compile Include="Models\Stubs\HttpStub.cs" />
     <Compile Include="Models\Imposters\HttpImposter.cs" />
     <Compile Include="Interfaces\IClient.cs" />
     <Compile Include="Models\Imposters\Imposter.cs" />
@@ -66,7 +66,7 @@
     <Compile Include="Models\Responses\Fields\TcpResponseFields.cs" />
     <Compile Include="Models\Responses\IsResponse.cs" />
     <Compile Include="Models\Responses\ResponseBase.cs" />
-    <Compile Include="Models\StubBase.cs" />
+    <Compile Include="Models\Stubs\StubBase.cs" />
     <Compile Include="MountebankClient.cs" />
     <Compile Include="Models\Predicates\EqualsPredicate.cs" />
     <Compile Include="MountebankRequestProxy.cs" />
